+++ lib/slimmer/skin.rb
@@ -22,10 +22,18 @@ module Slimmer
     def load_template(template_name)
       url = template_url(template_name)
       HTTPClient.get(url)
-    rescue RestClient::Exception => e
-      raise TemplateNotFoundException, "Unable to fetch: '#{template_name}' from '#{url}' because #{e}", caller
-    rescue Errno::ECONNREFUSED, SocketError, OpenSSL::SSL::SSLError => e
-      raise CouldNotRetrieveTemplate, "Unable to fetch: '#{template_name}' from '#{url}' because #{e}", caller
+    rescue Errno::ECONNREFUSED, SocketError, OpenSSL::SSL::SSLError, RestClient::Exception => e
+      message = "Unable to fetch: '#{template_name}' from '#{url}' because #{e}"
+
+      if e.is_a?(RestClient::Exception) && e.http_code == 404
+        raise TemplateNotFoundException, message, caller
+      end
+
+      if e.is_a?(RestClient::Exception) && [502, 503, 504].include?(e.http_code)
+        raise IntermittentRetrievalError, message, caller
+      end
+
+      raise CouldNotRetrieveTemplate, message, caller
     end
 
     def template_url(template_name)